extract models in common
