+++ app/src/main/java/org/shadowrunrussia2020/android/abilities/ActiveAbilitiesFragment.kt
@@ -10,10 +10,9 @@ import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import androidx.recyclerview.widget.LinearLayoutManager
 import kotlinx.android.synthetic.main.fragment_active_abilities.*
-import kotlinx.android.synthetic.main.fragment_spellbook.*
 import org.shadowrunrussia2020.android.R
 import org.shadowrunrussia2020.android.character.CharacterViewModel
-import org.shadowrunrussia2020.android.character.models.Character
+import org.shadowrunrussia2020.android.common.models.Character
 
 class ActiveAbilitiesFragment : Fragment() {
     private val mModel by lazy { ViewModelProviders.of(activity!!).get(CharacterViewModel::class.java) }
