+++ app/src/main/java/org/shadowrunrussia2020/android/billing/BillingViewModel.kt
@@ -4,8 +4,8 @@ import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import org.shadowrunrussia2020.android.ShadowrunRussia2020Application
-import org.shadowrunrussia2020.android.billing.models.AccountOverview
-import org.shadowrunrussia2020.android.billing.models.Transaction
+import org.shadowrunrussia2020.android.common.models.AccountOverview
+import org.shadowrunrussia2020.android.common.models.Transaction
 
 class BillingViewModel(application: Application) : AndroidViewModel(application) {
     private val mBillingRepository = BillingRepository(
