+++ app/src/main/java/org/shadowrunrussia2020/android/billing/BillingWebService.kt
@@ -1,9 +1,9 @@
 package org.shadowrunrussia2020.android.billing
 
 import kotlinx.coroutines.Deferred
-import org.shadowrunrussia2020.android.billing.models.AccountInfo
+import org.shadowrunrussia2020.android.common.models.AccountInfo
 import org.shadowrunrussia2020.android.models.Empty
-import org.shadowrunrussia2020.android.billing.models.Transfer
+import org.shadowrunrussia2020.android.common.models.Transfer
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
