--- app/src/main/java/org/shadowrunrussia2020/android/billing/models/AccountInfo.kt
@@ -1,3 +1,0 @@
-package org.shadowrunrussia2020.android.billing.models
-
-data  class AccountInfo(var balance: Int, var sin: Int, var history: List<Transaction>)
