+++ app/src/main/java/org/shadowrunrussia2020/android/di/IApplicationSingletonDi.kt
@@ -1,6 +1,6 @@
 package org.shadowrunrussia2020.android.di
 
 import org.shadowrunrussia2020.android.common.di.ApplicationSingletonScope
+import org.shadowrunrussia2020.android.main.ApplicationSingletonScopeMainDependency
 
-interface IApplicationSingletonDi: ApplicationSingletonScope {
-}
+interface IApplicationSingletonDi: ApplicationSingletonScope , ApplicationSingletonScopeMainDependency
