+++ app/src/main/java/org/shadowrunrussia2020/android/positioning/PositionsViewModel.kt
@@ -4,6 +4,7 @@ import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import org.shadowrunrussia2020.android.ShadowrunRussia2020Application
+import org.shadowrunrussia2020.android.common.models.Position
 
 class PositionsViewModel(application: Application) : AndroidViewModel(application) {
     private val mBillingRepository = PositionsRepository(
