+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -33,13 +33,13 @@
     <fragment android:id="@+id/transactionDetailsFragment"
               android:name="org.shadowrunrussia2020.android.billing.TransactionDetailsFragment"
               android:label="Подробности операции" tools:layout="@layout/fragment_transaction_details">
-        <argument android:name="transaction" app:argType="org.shadowrunrussia2020.android.billing.models.Transaction"/>
+        <argument android:name="transaction" app:argType="org.shadowrunrussia2020.android.common.models.Transaction"/>
     </fragment>
     <fragment android:id="@+id/spellDetailsFragment"
               android:name="org.shadowrunrussia2020.android.magic.SpellDetailsFragment"
               tools:layout="@layout/fragment_spell_details" android:label="Описание заклинания">
         <action android:id="@+id/action_choose_spell_target" app:destination="@id/scanQrFragmentSpell"/>
-        <argument android:name="spell" app:argType="org.shadowrunrussia2020.android.character.models.Spell"/>
+        <argument android:name="spell" app:argType="org.shadowrunrussia2020.android.common.models.Spell"/>
         <action
                 android:id="@+id/action_show_spell_result"
                 app:destination="@id/genericListFragment" />
@@ -135,7 +135,7 @@
             android:label="Описание способности">
         <argument
                 android:name="ability"
-                app:argType="org.shadowrunrussia2020.android.character.models.ActiveAbility" />
+                app:argType="org.shadowrunrussia2020.android.common.models.ActiveAbility" />
         <action
                 android:id="@+id/action_choose_ability_target"
                 app:destination="@id/scanQrFragmentAbility" />
@@ -146,7 +146,7 @@
             android:label="Описание способности">
         <argument
                 android:name="ability"
-                app:argType="org.shadowrunrussia2020.android.character.models.PassiveAbility" />
+                app:argType="org.shadowrunrussia2020.android.common.models.PassiveAbility" />
     </fragment>
     <fragment
             android:id="@+id/historyRecordDetailsFragment"
@@ -154,7 +154,7 @@
             android:label="Подробности" >
         <argument
                 android:name="record"
-                app:argType="org.shadowrunrussia2020.android.character.models.HistoryRecord" />
+                app:argType="org.shadowrunrussia2020.android.common.models.HistoryRecord" />
     </fragment>
     <fragment
             android:id="@+id/genericListFragment"
@@ -162,7 +162,7 @@
             android:label="Результат заклинания">
         <argument
                 android:name="records"
-                app:argType="org.shadowrunrussia2020.android.character.models.HistoryRecord[]" />
+                app:argType="org.shadowrunrussia2020.android.common.models.HistoryRecord[]" />
     </fragment>
     <fragment
             android:id="@+id/allPositionsFragment"
