+++ common/common/src/main/java/org/shadowrunrussia2020/android/common/declaration/PositionsDao.kt
@@ -1,10 +1,11 @@
-package org.shadowrunrussia2020.android.positioning
+package org.shadowrunrussia2020.android.common.declaration
 
 import androidx.lifecycle.LiveData
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
+import org.shadowrunrussia2020.android.common.models.Position
 
 @Dao
 interface PositionsDao {
