--- features/main_screen/src/main/java/org/shadowrunrussia2020/android/main/ActivityScopeMainDependency.kt
@@ -1,7 +1,0 @@
-package org.shadowrunrussia2020.android.main
-
-import org.shadowrunrussia2020.android.common.di.MainActivityScope
-
-interface ActivityScopeMainDependency : MainActivityScope {
-
-}
