+++ features/main_screen/src/main/java/org/shadowrunrussia2020/android/main/Dependency.kt
@@ -1,0 +1,8 @@
+package org.shadowrunrussia2020.android.main
+
+import org.shadowrunrussia2020.android.common.di.ApplicationSingletonScope
+import org.shadowrunrussia2020.android.common.di.MainActivityScope
+
+interface ActivityScopeMainDependency : MainActivityScope
+
+interface ApplicationSingletonScopeMainDependency: ApplicationSingletonScope
