+++ features/main_screen/src/main/java/org/shadowrunrussia2020/android/main/MainViewModel.kt
@@ -1,0 +1,14 @@
+package org.shadowrunrussia2020.android.main
+
+import android.app.Application
+import androidx.lifecycle.AndroidViewModel
+import java.lang.RuntimeException
+
+internal class MainViewModel(application: Application) : AndroidViewModel(application) {
+
+    val dependency = application as? ApplicationSingletonScopeMainDependency
+        ?: RuntimeException("ApplicationSingletonScopeMAinDependency must be implement on application")
+
+
+
+}
